enhance in repositories to be in StatesRepository
