+++ src/main/kotlin/planmate/data/repositoryImp/StatesRepositoryImp.kt
@@ -2,13 +2,20 @@ package planmate.data.repositoryImp
 
 import planmate.data.csvHandler.CsvFileHandler
 import planmate.domain.models.State
-import planmate.domain.repository.CreateStateRepository
+import planmate.domain.repository.StatesRepository
 import kotlin.uuid.ExperimentalUuidApi
+import kotlin.uuid.Uuid
 
-class CreateStateRepositoryImp(
+class StatesRepositoryImp(
     private val csvFileHandler: CsvFileHandler,
-) : CreateStateRepository {
+) : StatesRepository {
+
     private val header = arrayOf("id", "name", "projectId")
+
+    @OptIn(ExperimentalUuidApi::class)
+    override fun getAllStates(projectId: Uuid): List<State> {
+        TODO("Not yet implemented")
+    }
 
     @OptIn(ExperimentalUuidApi::class)
     override fun createState(state: State) {
@@ -17,5 +24,14 @@ class CreateStateRepositoryImp(
         val stateRow = arrayOf(state.id.toString(), state.name, state.projectId.toString())
 
         csvFileHandler.appendLine(filename = filename, headerColumns = header, newRow = stateRow)
+    }
+
+    override fun updateState(state: State) {
+        TODO("Not yet implemented")
+    }
+
+    @OptIn(ExperimentalUuidApi::class)
+    override fun deleteState(projectId: Uuid) {
+        TODO("Not yet implemented")
     }
 }
