--- src/main/kotlin/planmate/data/repositoryImp/UpdateStateRepositoryImp.kt
@@ -1,13 +1,0 @@
-package planmate.data.repositoryImp
-
-import planmate.data.csvHandler.CsvFileHandler
-import planmate.domain.models.State
-import planmate.domain.repository.UpdateStateRepository
-
-class UpdateStateRepositoryImp(
-    private val csvFileHandler: CsvFileHandler,
-) : UpdateStateRepository {
-    override fun updateState(state: State) {
-        TODO("Not yet implemented")
-    }
-}
