--- src/main/kotlin/planmate/domain/repository/CreateStateRepository.kt
@@ -1,7 +1,0 @@
-package planmate.domain.repository
-
-import planmate.domain.models.State
-
-interface CreateStateRepository {
-    fun createState(state: State)
-}
