--- src/main/kotlin/planmate/domain/repository/DeleteStateRepository.kt
@@ -1,9 +1,0 @@
-package planmate.domain.repository
-
-import kotlin.uuid.ExperimentalUuidApi
-import kotlin.uuid.Uuid
-
-interface DeleteStateRepository {
-    @OptIn(ExperimentalUuidApi::class)
-    fun deleteState( projectId: Uuid)
-}
