+++ src/main/kotlin/planmate/domain/repository/StatesRepository.kt
@@ -4,7 +4,15 @@ import planmate.domain.models.State
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
 
-interface GetAllStatesRepository {
+interface StatesRepository {
+
     @OptIn(ExperimentalUuidApi::class)
     fun getAllStates(projectId: Uuid): List<State>
+
+    fun createState(state: State)
+
+    fun updateState(state: State)
+
+    @OptIn(ExperimentalUuidApi::class)
+    fun deleteState(projectId: Uuid)
 }
