--- src/main/kotlin/planmate/domain/repository/UpdateStateRepository.kt
@@ -1,7 +1,0 @@
-package planmate.domain.repository
-
-import planmate.domain.models.State
-
-interface UpdateStateRepository {
-    fun updateState(state: State)
-}
